pyproject.toml
[tool.poetry]
name = "sisense-sync"
version = "1.0.0"
description = ""
authors = ["Chris Alexander <crainte@hybridoptix.net>"]

[tool.poetry.dependencies]
python = "^3.9"
GitPython = "^3.1.14"
loguru = "^0.5.3"
PySenseSDK = "^1.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
sisense = "sisense_sync.actions:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
